requirements.txt
requests
beautifulsoup4
numpy==1.17.2
gensim==3.7.1
pandas==0.24.2
pdfminer.six==20181108
spacy==2.1.3
PyYAML==5.1
ruamel.yaml==0.15.35